pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0a1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "markdown>=3.9",
    "psycopg[binary,pool]>=3.2.10",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest-django>=4.11.1",
    "ruff>=0.13.1",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
manage = "python src/manage.py"
server = "python src/manage.py runserver"
migrate = "python src/manage.py migrate"
makemigrations = "python src/manage.py makemigrations"

wsgi = "gunicorn backend.wsgi:application --bind 0.0.0.0:8000"

format = "ruff format src tests"
check = "ruff check src tests"

test = "pytest"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]
python_files = ["test_*.py"]

DJANGO_SETTINGS_MODULE = "config.settings"